Introduce round and character objects.
A round is a battle royale between the players on a given map (in a given arena)
A round is finished when one or no player is remaining.

A character represents the identity of a player.
It is stable accross a game.
Players are created for each round.

Make the bullet registry a class.
Each round instantiates a local bullet registry.

A map has starting positions to dispatch the players at the beginning of a round.

Update main.lua, it creates characters, a map and then instantiate a round.