Introduce game states and a game state scheduler.
The application is a graph of game states.
A game state scheduler handles transitions between game states.
Introduce three first game states:
- WelcomeGameState: the welcome screen
- RoundGameState: a round
- ScoreGameState: display the current score of a game
At startup the application displays the welcome screen, then a game is started with a round limit of three.
Between each round the score is displayed. When the game is finished (after three rounds), we come back to the welcome screen.